litellm_arize/requirements.txt
litellm
prometheus_client
arize-phoenix
locust
openai
arize-phoenix-client 
openinference-instrumentation-openai 
datasets 
duckdb 
pandas
httpx<0.28